+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * index: Fixed crash with `RuntimeError` after continuous realtime connection loss.
 
+### Performance
+
+* cli: Lazy import dependencies to speed up startup.
+
 ## 5.1.5 - 2022-06-05
 
 ### Fixed