+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## [unreleased]
 
+### Fixed
+
+* cli: Commands with `--help` option no longer require a working DipDup config.
+
 ### Performance
 
 * cli: Lazy import dependencies to speed up startup.
+
+### Other
+
+* docs: Migrate docs from GitBook to mdbook.
 
 ## 5.1.5 - 2022-06-05